+++ Eplant/ResponsiveDrawer.tsx
@@ -1,6 +1,7 @@
 
 import {Drawer, DrawerProps} from '@mui/material'
 import {useState} from 'react'
+// TODO: Make this drawer support opening/closing on mobile
 
 const ResponsiveDrawer = (props: DrawerProps) => {
   const [open, setOpen] = useState(props.open);
